Rename Refitter namespace
