+++ src/VSIX/ApiClientCodeGen.VSIX.Shared/Generators/CodeGeneratorFactory.cs
@@ -6,7 +6,7 @@ using Rapicgen.Core.Generators.AutoRest;
 using Rapicgen.Core.Generators.Kiota;
 using Rapicgen.Core.Generators.NSwag;
 using Rapicgen.Core.Generators.OpenApi;
-using Rapicgen.Core.Generators.Refit;
+using Rapicgen.Core.Generators.Refitter;
 using Rapicgen.Core.Generators.Swagger;
 using Rapicgen.Core.Installer;
 using Rapicgen.Core.Logging;
@@ -103,7 +103,7 @@ namespace Rapicgen.Generators
                         dependencyInstaller);
                 
                 case SupportedCodeGenerator.Refitter:
-                    return new RefitCodeGenerator(inputFilePath, defaultNamespace);
+                    return new RefitterCodeGenerator(inputFilePath, defaultNamespace);
 
                 default:
                     throw new NotSupportedException();
